+++ src/components/CardExpirationSection/CardExpirationSection.tsx
@@ -1,5 +1,5 @@
 import styles from './CardExpirationSection.module.css';
-import { InputSection } from '../common/InputSection/InputSection';
+import { FieldGroup } from '../common/FieldGroup/FieldGroup';
 import { Dispatch, SetStateAction, useState } from 'react';
 import {
   validateMonthRangeError,
@@ -7,6 +7,7 @@ import {
   validateYearLengthError
 } from '../../utils/CardInputValidations';
 import { Expiration } from '../../types/card';
+import { InputWrapper } from '../common/InputWrapper/InputWrapper';
 
 type Props = {
   expiration: Expiration;
@@ -32,13 +33,13 @@ export default function CardExpirationSection({ expiration, setExpiration }: Pro
 
   return (
     <div className={styles.sectionContainer}>
-      <InputSection.TitleWrapper>
-        <InputSection.Title title="카드 유효기간을 입력해 주세요" />
-        <InputSection.SubTitle title="월/년도(MMYY)를 순서대로 입력해 주세요." />
-      </InputSection.TitleWrapper>
-      <div className={styles.inputSection}>
-        <InputSection.Label text="유효기간" />
-        <InputSection.InputWrapper<'month' | 'year'>
+      <FieldGroup.TitleWrapper>
+        <FieldGroup.Title title="카드 유효기간을 입력해 주세요" />
+        <FieldGroup.SubTitle title="월/년도(MMYY)를 순서대로 입력해 주세요." />
+      </FieldGroup.TitleWrapper>
+      <div className={styles.FieldGroup}>
+        <FieldGroup.Label text="유효기간" />
+        <InputWrapper<'month' | 'year'>
           fields={[
             { key: 'month', value: expiration.month },
             { key: 'year', value: expiration.year }
@@ -49,8 +50,8 @@ export default function CardExpirationSection({ expiration, setExpiration }: Pro
           maxLength={2}
         />
         <div>
-          {expirationError.month && <InputSection.Error message={expirationError.month} />}
-          {expirationError.year && <InputSection.Error message={expirationError.year} />}
+          {expirationError.month && <FieldGroup.Error message={expirationError.month} />}
+          {expirationError.year && <FieldGroup.Error message={expirationError.year} />}
         </div>
       </div>
     </div>
