+++ src/components/CardNumberSection/CardNumberSection.module.css
@@ -3,6 +3,6 @@
   gap: 10px;
 }
 
-.inputSection {
+.FieldGroup {
   height: 77px;
 }
