+++ src/components/CvcSection/CvcSection.module.css
@@ -3,6 +3,6 @@
   gap: 10px;
 }
 
-.inputSection {
+.FieldGroup {
   height: 77px;
 }
