+++ src/components/common/InputWrapper/InputWrapper.module.css
@@ -2,7 +2,3 @@
   display: flex;
   gap: 10px;
 }
-
-.FieldGroup {
-  height: 77px;
-}
