+++ src/components/common/InputWrapper/InputWrapper.tsx
@@ -1,26 +1,5 @@
-import styles from './InputSection.module.css';
+import styles from './InputWrapper.module.css';
 import Input from '../Input/Input';
-import { ReactNode } from 'react';
-
-function Title({ title }: { title: string }) {
-  return <h1>{title}</h1>;
-}
-
-function SubTitle({ title }: { title: string }) {
-  return <p>{title}</p>;
-}
-
-function TitleWrapper({ children }: { children: ReactNode }) {
-  return <div className={styles.titleWrapper}>{children}</div>;
-}
-
-function Label({ text }: { text: string }) {
-  return <label className={styles.inputTitle}>{text}</label>;
-}
-
-function Error({ message }: { message: string }) {
-  return <p className={styles.errorMessage}>{message}</p>;
-}
 
 type InputField<T extends string = string> = {
   key: T;
@@ -57,12 +36,3 @@ export function InputWrapper<T extends string = string>({
     </div>
   );
 }
-
-export const InputSection = {
-  TitleWrapper,
-  Title,
-  SubTitle,
-  Error,
-  InputWrapper,
-  Label
-};
